Refactoring routes
