+++ app/routes/__init__.py
@@ -1,0 +1,7 @@
+from flask import Blueprint
+routes = Blueprint('routes', __name__)
+
+from .root import *
+from .courses import *
+from .periods import *
+from .triples import *
